requirements.txt
requests
python-telegram-bot
Jinja2